Add resume, medium, email to
