+++ src/components/Home.tsx
@@ -1,15 +1,18 @@
 import { useEffect, useState } from 'react';
-import info from './info.json';
+import info from '../info.json';
+import SocialMediaLink from './SocialMediaLink';
 
 interface IInfo {
   fullname: string;
   title: string;
   location: string;
   description: string;
+  email: string;
+  resume: string;
   social: ISocialMediaLink[];
 }
 
-interface ISocialMediaLink {
+export interface ISocialMediaLink {
   href: string;
   title: string;
   username: string;
@@ -17,27 +20,6 @@ interface ISocialMediaLink {
 }
 
 const personalInfo: IInfo = info;
-
-const SocialMediaLink = ({
-  href,
-  title,
-  svgPath,
-  username,
-}: ISocialMediaLink) => {
-  return (
-    <a className="link" href={href} data-tippy-content={username}>
-      <svg
-        className="h-6 fill-current text-gray-600 dark:text-gray-100 hover:text-green-700"
-        role="img"
-        viewBox="0 0 24 24"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <title>{title}</title>
-        <path d={svgPath} />
-      </svg>
-    </a>
-  );
-};
 
 const Home = () => {
   const [darkMode, setDarkMode] = useState<boolean>(true);
@@ -84,9 +66,22 @@ const Home = () => {
           </p>
 
           <div className="pt-12 pb-8">
-            <button className="bg-green-700 hover:bg-green-900 text-white font-bold py-2 px-4 rounded-full">
+            <a
+              className="bg-green-700 hover:bg-green-900 text-white font-bold py-4 px-6 rounded-full"
+              href={`mailto:${personalInfo.email}`}
+              target="_blank"
+              rel="noreferrer"
+            >
               Get In Touch
-            </button>
+            </a>
+            <a
+              className="text-green-700 dark:text-white underline font-bold py-4 px-6"
+              href={personalInfo.resume}
+              target="_blank"
+              rel="noreferrer"
+            >
+              Resume
+            </a>
           </div>
 
           <div className="mt-6 pb-16 lg:pb-0 w-4/5 lg:w-full mx-auto flex flex-wrap items-center justify-between">
@@ -101,17 +96,17 @@ const Home = () => {
       </div>
       <div className="fixed top-2 right-2">
         <button
-          className="rounded-full bg-green-700 hover:bg-green-900 hover:bg-grey text-grey-darkest font-bold p-2 inline-flex items-center opacity-75 dark:opacity-75 shadow"
+          className="rounded-full bg-gray-100 dark:bg-gray-700 hover:bg-white dark:hover:bg-gray-900 hover:bg-grey text-grey-darkest font-bold p-2 inline-flex items-center opacity-75 dark:opacity-75 shadow"
           onClick={() => setDarkMode(!darkMode)}
         >
           <svg
-            className="w-8 h-8 fill-white"
+            className="w-8 h-8 fill-gray-900 dark:fill-gray-100"
             xmlns="http://www.w3.org/2000/svg"
             viewBox="0 0 24 24"
           >
             <path
               d={
-                darkMode
+                !darkMode
                   ? 'M12.048 21.963c-.308 0-.618-.015-.93-.043-2.66-.246-5.064-1.513-6.771-3.567s-2.512-4.651-2.266-7.311a10.004 10.004 0 0 1 9.038-9.038 1 1 0 0 1 .896 1.589 6.008 6.008 0 0 0 1.258 8.392c2.078 1.536 5.055 1.536 7.133 0a1 1 0 0 1 1.591.896 9.951 9.951 0 0 1-9.949 9.082zM9.315 4.438a8.006 8.006 0 0 0-5.244 6.787 7.954 7.954 0 0 0 1.813 5.849 7.95 7.95 0 0 0 5.417 2.854 7.95 7.95 0 0 0 8.266-5.243 8.01 8.01 0 0 1-2.729.476 7.946 7.946 0 0 1-4.755-1.565C9.174 11.443 8.145 7.68 9.315 4.438z'
                   : 'M12 18c-3.309 0-6-2.691-6-6s2.691-6 6-6 6 2.691 6 6-2.691 6-6 6zm0-10c-2.206 0-4 1.794-4 4s1.794 4 4 4 4-1.794 4-4-1.794-4-4-4zm0-4a1 1 0 0 1-1-1V1a1 1 0 0 1 2 0v2a1 1 0 0 1-1 1zm0 20a1 1 0 0 1-1-1v-2a1 1 0 1 1 2 0v2a1 1 0 0 1-1 1zM5.64 6.64a.997.997 0 0 1-.707-.293l-1.42-1.42a.999.999 0 1 1 1.414-1.414l1.42 1.42A.999.999 0 0 1 5.64 6.64zm14.139 14.139a.997.997 0 0 1-.707-.293l-1.42-1.42a.999.999 0 1 1 1.414-1.414l1.42 1.42a.999.999 0 0 1-.707 1.707zM3 13H1a1 1 0 1 1 0-2h2a1 1 0 0 1 0 2zm20 0h-2a1 1 0 1 1 0-2h2a1 1 0 1 1 0 2zM4.22 20.779a.999.999 0 0 1-.707-1.707l1.42-1.42a.999.999 0 1 1 1.414 1.414l-1.42 1.42a.993.993 0 0 1-.707.293zM18.359 6.64a.999.999 0 0 1-.707-1.707l1.42-1.42a.999.999 0 1 1 1.414 1.414l-1.42 1.42a.997.997 0 0 1-.707.293z'
               }
